# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Self-Healing Test Automation Harness API
  description: |
    A comprehensive REST API for managing test execution, monitoring results, and configuring self-healing capabilities.
    
    ## Features
    - **Test Orchestration**: Execute and manage multiple types of tests (unit, e2e, performance, security)
    - **Self-Healing**: AI-powered test recovery and adaptation
    - **Unified Reporting**: Comprehensive test results and analytics
    - **Plugin Architecture**: Extensible test engine support
    - **Real-time Monitoring**: Live test execution status and progress tracking
    
    ## Authentication
    Currently, the API operates without authentication for development purposes. In production, API key authentication will be implemented.
  version: 1.0.0
  contact:
    name: Test Automation Harness Support
    email: support@test-automation-harness.com
    url: https://github.com/your-org/test-automation-harness
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.test-automation-harness.com
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/healing/strategies:
    get:
      summary: Get available healing strategies
      description: Get list of available healing strategies
      tags:
        - Healing Management
      parameters:
        - name: failureType
          in: query
          description: Filter by failure type
          schema:
            type: string
        - name: enabled
          in: query
          description: Filter by enabled status
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Available healing strategies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  - id: "css-fallback"
                    name: "CSS Fallback Strategy"
                    description: "Falls back to alternative CSS selectors"
                    failureTypes: ["element_not_found", "timeout"]
                    enabled: true
                    confidence: 0.8
                    config: {}
  /api/v1/healing/attempts:
    get:
      summary: Get healing attempts with filtering and pagination
      description: Get healing attempts with various filtering and pagination options
      tags:
        - Healing Management
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/SortByParam'
        - name: testId
          in: query
          description: Filter by test ID
          schema:
            type: string
        - name: strategy
          in: query
          description: Filter by strategy name
          schema:
            type: string
        - name: success
          in: query
          description: Filter by success status
          schema:
            type: boolean
        - name: minConfidence
          in: query
          description: Minimum confidence threshold
          schema:
            type: number
            minimum: 0
            maximum: 1
        - name: startDate
          in: query
          description: Filter by start date
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter by end date
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Healing attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  attempts:
                    - id: "attempt_123"
                      testId: "test_123"
                      strategy: "css-fallback"
                      failureType: "element_not_found"
                      success: true
                      confidence: 0.8
                      duration: 1500
                      timestamp: "2025-01-06T15:30:00.000Z"
                      changes:
                        selector: "button.submit"
                        newSelector: "input[type='submit']"
                  pagination:
                    page: 1
                    limit: 10
                    total: 50
                    totalPages: 5
                    hasNext: true
                    hasPrev: false
  /api/v1/healing/statistics:
    get:
      summary: Get healing statistics and metrics
      description: Get comprehensive healing statistics and metrics
      tags:
        - Healing Management
      parameters:
        - name: startDate
          in: query
          description: Filter by start date
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter by end date
          schema:
            type: string
            format: date-time
        - name: groupBy
          in: query
          description: Group by field
          schema:
            type: string
            enum: [strategy, failureType, testId]
        - name: minConfidence
          in: query
          description: Minimum confidence threshold
          schema:
            type: number
            minimum: 0
            maximum: 1
      responses:
        '200':
          description: Healing statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  totalAttempts: 1000
                  successfulAttempts: 800
                  failedAttempts: 200
                  successRate: 0.8
                  avgConfidence: 0.75
                  avgDuration: 2000
                  strategies:
                    - name: "css-fallback"
                      attempts: 500
                      successful: 400
                      successRate: 0.8
                      avgConfidence: 0.75
                  byFailureType:
                    element_not_found:
                      attempts: 600
                      successful: 480
                    timeout:
                      attempts: 400
                      successful: 320
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (currently optional in development)
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortParam:
      name: sort
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    SortByParam:
      name: sortBy
      in: query
      description: Sort field
      schema:
        type: string
        default: createdAt
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: "2025-01-06T15:30:00.000Z"
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
...